+++ ppci/lang/pascal/__init__.py
@@ -1,0 +1,7 @@
+
+""" Pascal front-end """
+
+from .builder import PascalBuilder
+
+
+__all__ = ['PascalBuilder']